Improved compatibility with HS2_StudioWindowResize
An overlapping issue occurred because Studio_AnimationSearch could not adapt to the vertically increased length by HS2_StudioWindowResize

I tried to port this to KK, but the translate stuff in this code didn't work well in KK and it felt very heavy. Maybe it needs caching stuff 
Also, the nodes of AnimeGroupList and AnimeCategoryList were not filtered. I assume it was because the InitList was different from the ones in HS2/AI

Anyway, I made a PR for something I can improve right now